# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "StudentPlanner.UI",
    "version": "1.0"
  },
  "paths": {
...
    "/events/personal-events": {
      "get": {
        "tags": [
          "PersonalEvents"
        ],
        "summary": "Retrieves all personal events belonging to the authenticated user.",
        "description": "This endpoint returns only the events owned by the currently logged-in user.\r\nThe user identifier is derived from the authentication context and not\r\nfrom client input.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "You don't have permission to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Returns the list of personal events.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalEventResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "PersonalEvents"
        ],
        "summary": "Creates a new personal event for the authenticated user.",
        "description": "The created event is automatically associated with the current user.\r\nClients do not provide user identifiers directly.",
        "requestBody": {
          "description": "The event creation payload.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalEventRequest"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "You don't have permission to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "201": {
            "description": "Returns the newly created personal event."
          }
        }
      }
    },
    "/events/personal-events/{id}": {
      "get": {
        "tags": [
          "PersonalEvents"
        ],
        "summary": "Retrieves all personal events belonging to the authenticated user.",
        "description": "This endpoint returns only the events owned by the currently logged-in user.\r\nThe user identifier is derived from the authentication context and not\r\nfrom client input.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "You don't have permission to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Returns the list of personal events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "PersonalEvents"
        ],
        "summary": "Updates an existing personal event.",
        "description": "The authenticated user may only update events they own.\r\nAttempts to update events owned by other users will result in a forbidden response.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated event data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalEventUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalEventUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "If the event does not belong to the authenticated user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Returns the updated event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "If the event identifier does not match the payload."
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "If the event does not exist."
          }
        }
      },
      "delete": {
        "tags": [
          "PersonalEvents"
        ],
        "summary": "Deletes an existing personal event.",
        "description": "The authenticated user may only delete events they own.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the event.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "If the event does not belong to the authenticated user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "Event deleted successfully."
          },
          "404": {
            "description": "If the event does not exist."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatePersonalEventRequest": {
        "type": "object",
        "properties": {
          "details": {
            "$ref": "#/components/schemas/EventDetailsInputDTO"
          }
        },
        "additionalProperties": false
      },
...
      "EventDetailsInputDTO": {
        "required": [
          "endTime",
          "eventId",
          "location",
          "startTime",
          "title"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventDetailsOutputDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PersonalEventResponse": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "$ref": "#/components/schemas/EventDetailsOutputDTO"
          }
        },
        "additionalProperties": false
      },
      "PersonalEventUpdateDTO": {
        "type": "object",
        "properties": {
          "details": {
            "$ref": "#/components/schemas/EventDetailsInputDTO"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}